Move `create_sample_table`, `sample_to_chunk_iter`, `media_time_to_us`, `track_time_to_us` `Mp4parseIndice` and `Mp4parseTrackType` to `mp4parse` crate, as they are not part of the exposed C API and can be also useful from Rust
