+++ mp4parse_capi/tests/test_sample_table.rs
@@ -1,5 +1,8 @@
 extern crate mp4parse_capi;
+extern crate mp4parse;
 use mp4parse_capi::*;
+use mp4parse::TrackType;
+use mp4parse::Indice;
 use std::io::Read;
 
 extern "C" fn buf_read(buf: *mut u8, size: usize, userdata: *mut std::os::raw::c_void) -> isize {
@@ -34,7 +37,7 @@ fn parse_sample_table() {
         let mut track_info = Mp4parseTrackInfo::default();
         rv = mp4parse_get_track_info(parser, 1, &mut track_info);
         assert_eq!(rv, Mp4parseStatus::Ok);
-        assert_eq!(track_info.track_type, Mp4parseTrackType::Audio);
+        assert_eq!(track_info.track_type, TrackType::Audio);
 
         // Check audio smaple table
         let mut is_fragmented_file: u8 = 0;
@@ -47,7 +50,7 @@ fn parse_sample_table() {
         assert_eq!(rv, Mp4parseStatus::Ok);
 
         // Compare the value from stagefright.
-        let audio_indice_0 = Mp4parseIndice {
+        let audio_indice_0 = Indice {
             start_offset: 27_046.into(),
             end_offset: 27_052.into(),
             start_composition: 0.into(),
@@ -55,7 +58,7 @@ fn parse_sample_table() {
             start_decode: 0.into(),
             sync: true,
         };
-        let audio_indice_215 = Mp4parseIndice {
+        let audio_indice_215 = Indice {
             start_offset: 283_550.into(),
             end_offset: 283_556.into(),
             start_composition: 9_984_580.into(),
@@ -70,7 +73,7 @@ fn parse_sample_table() {
         // Check video smaple table
         rv = mp4parse_get_track_info(parser, 0, &mut track_info);
         assert_eq!(rv, Mp4parseStatus::Ok);
-        assert_eq!(track_info.track_type, Mp4parseTrackType::Video);
+        assert_eq!(track_info.track_type, TrackType::Video);
 
         let mut is_fragmented_file: u8 = 0;
         rv = mp4parse_is_fragmented(parser, track_info.track_id, &mut is_fragmented_file);
@@ -82,7 +85,7 @@ fn parse_sample_table() {
         assert_eq!(rv, Mp4parseStatus::Ok);
 
         // Compare the last few data from stagefright.
-        let video_indice_291 = Mp4parseIndice {
+        let video_indice_291 = Indice {
             start_offset: 280_226.into(),
             end_offset: 280_855.into(),
             start_composition: 9_838_333.into(),
@@ -90,7 +93,7 @@ fn parse_sample_table() {
             start_decode: 9_710_000.into(),
             sync: false,
         };
-        let video_indice_292 = Mp4parseIndice {
+        let video_indice_292 = Indice {
             start_offset: 280_855.into(),
             end_offset: 281_297.into(),
             start_composition: 9_805_011.into(),
@@ -100,9 +103,9 @@ fn parse_sample_table() {
         };
         // TODO: start_composition time in stagefright is 9905000, but it is 9904999 in parser, it
         //       could be rounding error.
-        //let video_indice_293 = Mp4parseIndice { start_offset: 281_297, end_offset: 281_919, start_composition: 9_905_000, end_composition: 9_938_344, start_decode: 9_776_666, sync: false };
-        //let video_indice_294 = Mp4parseIndice { start_offset: 281_919, end_offset: 282_391, start_composition: 9_871_677, end_composition: 9_905_000, start_decode: 9_776_677, sync: false };
-        let video_indice_295 = Mp4parseIndice {
+        //let video_indice_293 = Indice { start_offset: 281_297, end_offset: 281_919, start_composition: 9_905_000, end_composition: 9_938_344, start_decode: 9_776_666, sync: false };
+        //let video_indice_294 = Indice { start_offset: 281_919, end_offset: 282_391, start_composition: 9_871_677, end_composition: 9_905_000, start_decode: 9_776_677, sync: false };
+        let video_indice_295 = Indice {
             start_offset: 282_391.into(),
             end_offset: 283_032.into(),
             start_composition: 9_971_666.into(),
@@ -110,7 +113,7 @@ fn parse_sample_table() {
             start_decode: 9_843_333.into(),
             sync: false,
         };
-        let video_indice_296 = Mp4parseIndice {
+        let video_indice_296 = Indice {
             start_offset: 283_092.into(),
             end_offset: 283_526.into(),
             start_composition: 9_938_344.into(),
@@ -153,7 +156,7 @@ fn parse_sample_table_with_elst() {
         let mut track_info = Mp4parseTrackInfo::default();
         rv = mp4parse_get_track_info(parser, 1, &mut track_info);
         assert_eq!(rv, Mp4parseStatus::Ok);
-        assert_eq!(track_info.track_type, Mp4parseTrackType::Audio);
+        assert_eq!(track_info.track_type, TrackType::Audio);
 
         // Check audio sample table
         let mut is_fragmented_file: u8 = std::u8::MAX;
@@ -168,7 +171,7 @@ fn parse_sample_table_with_elst() {
         // Compare the value from stagefright.
         // Due to 'elst', the start_composition and end_composition are negative
         // at first two samples.
-        let audio_indice_0 = Mp4parseIndice {
+        let audio_indice_0 = Indice {
             start_offset: 6992.into(),
             end_offset: 7363.into(),
             start_composition: (-36281).into(),
@@ -176,7 +179,7 @@ fn parse_sample_table_with_elst() {
             start_decode: 0.into(),
             sync: true,
         };
-        let audio_indice_1 = Mp4parseIndice {
+        let audio_indice_1 = Indice {
             start_offset: 7363.into(),
             end_offset: 7735.into(),
             start_composition: (-13062).into(),
@@ -184,7 +187,7 @@ fn parse_sample_table_with_elst() {
             start_decode: 23219.into(),
             sync: true,
         };
-        let audio_indice_2 = Mp4parseIndice {
+        let audio_indice_2 = Indice {
             start_offset: 7735.into(),
             end_offset: 8106.into(),
             start_composition: 10158.into(),
@@ -223,7 +226,7 @@ fn parse_sample_table_with_negative_ctts() {
         let mut track_info = Mp4parseTrackInfo::default();
         rv = mp4parse_get_track_info(parser, 0, &mut track_info);
         assert_eq!(rv, Mp4parseStatus::Ok);
-        assert_eq!(track_info.track_type, Mp4parseTrackType::Video);
+        assert_eq!(track_info.track_type, TrackType::Video);
 
         let mut is_fragmented_file: u8 = std::u8::MAX;
         rv = mp4parse_is_fragmented(parser, track_info.track_id, &mut is_fragmented_file);
@@ -235,7 +238,7 @@ fn parse_sample_table_with_negative_ctts() {
         assert_eq!(rv, Mp4parseStatus::Ok);
 
         // There are negative value in 'ctts' table.
-        let video_indice_0 = Mp4parseIndice {
+        let video_indice_0 = Indice {
             start_offset: 48.into(),
             end_offset: 890.into(),
             start_composition: 0.into(),
@@ -243,7 +246,7 @@ fn parse_sample_table_with_negative_ctts() {
             start_decode: 0.into(),
             sync: true,
         };
-        let video_indice_1 = Mp4parseIndice {
+        let video_indice_1 = Indice {
             start_offset: 890.into(),
             end_offset: 913.into(),
             start_composition: 133_333.into(),
@@ -251,7 +254,7 @@ fn parse_sample_table_with_negative_ctts() {
             start_decode: 33_333.into(),
             sync: false,
         };
-        let video_indice_2 = Mp4parseIndice {
+        let video_indice_2 = Indice {
             start_offset: 913.into(),
             end_offset: 934.into(),
             start_composition: 66_666.into(),
@@ -259,7 +262,7 @@ fn parse_sample_table_with_negative_ctts() {
             start_decode: 66_666.into(),
             sync: false,
         };
-        let video_indice_3 = Mp4parseIndice {
+        let video_indice_3 = Indice {
             start_offset: 934.into(),
             end_offset: 955.into(),
             start_composition: 33_333.into(),
